% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_color}
\alias{gfx_color}
\title{Sets the current color for drawing}
\usage{
gfx_color(red, green, blue)
}
\arguments{
\item{red}{Integer. A value between 0 and 255.}

\item{green}{Integer. A value between 0 and 255.}

\item{blue}{Integer. A value between 0 and 255.}
}
\description{
The drawing color is set using the RGB color model, in which you mix red,
green, and blue components in order to get the desired color.
}
