% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_event_waiting}
\alias{gfx_event_waiting}
\title{Check if an event has occurred without blocking}
\usage{
gfx_event_waiting()
}
\value{
Logical. TRUE if an event occurred. FALSE otherwise.
}
\description{
\code{\link[=gfx_wait]{gfx_wait()}} causes the program to wait until something is pressed, which
can be inconvenient in a real-time program. To avoid this, call this
function, which returns \code{TRUE} if an event (key or button press) has
occurred, otherwise returns \code{FALSE} immediately. If it returns \code{TRUE},
then you can call \code{\link[=gfx_wait]{gfx_wait()}} to retrieve the event without waiting.
}
