% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_flush}
\alias{gfx_flush}
\title{Flush all output to the graphics window}
\usage{
gfx_flush()
}
\description{
This forces all previous drawing commands to take effect. Output is normally
flushed after a call to \code{\link[=gfx_wait]{gfx_wait()}}. If your program does not call
\code{\link[=gfx_wait]{gfx_wait()}}, then you need to call \code{gfx_flush} after drawing to the screen.
}
