% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_line}
\alias{gfx_line}
\title{Draw a line in the graphics window using the current color}
\usage{
gfx_line(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Integer. From this X position.}

\item{y1}{Integer. From this Y position.}

\item{x2}{Integer. To this X position.}

\item{y2}{Integer. To this Y position.}
}
\description{
Draw a line in the graphics window using the current color
}
