% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_open}
\alias{gfx_open}
\title{Create a new graphics window on the screen}
\usage{
gfx_open(width, height, title = "gfxR")
}
\arguments{
\item{width}{Integer. Width of the window in pixels.}

\item{height}{Integer. Height of the window in pixels.}

\item{title}{String. Name of the window in the title bar.}
}
\description{
Create a new graphics window on the screen
}
