% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_point}
\alias{gfx_point}
\title{Draw a single point at (x, y)}
\usage{
gfx_point(x, y)
}
\arguments{
\item{x}{Integer. Position X on the window.}

\item{y}{Integer. Position Y on the window.}
}
\description{
Draw a single point at (x, y)
}
