% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_wait}
\alias{gfx_wait}
\title{Wait until the user press a key or a mouse button}
\usage{
gfx_wait()
}
\value{
Either an integer with the mouse button number pressed,
or the keyboard character pressed.
}
\description{
If the user has pressed a mouse button, the integer 1, 2, or 3, will be
returned, indicating the button pressed. If the user has pressed a key,
that character will be returned. To determine the mouse location at that
time, call [gfx_xpos() and \code{\link[=gfx_ypos]{gfx_ypos()}}].
}
