% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_xpos}
\alias{gfx_xpos}
\title{Return the pressed X coordinate}
\usage{
gfx_xpos()
}
\value{
Integer with the X coordinate
}
\description{
\code{gfx_xpos()} and \code{\link[=gfx_ypos]{gfx_ypos()}} return the X and Y coordinates, respectively,
of the mouse pointer when the last key or button was pressed.
}
