% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfxfuns.R
\name{gfx_ypos}
\alias{gfx_ypos}
\title{Return the pressed Y coordinate}
\usage{
gfx_ypos()
}
\value{
Integer with the Y coordinate
}
\description{
\code{\link[=gfx_xpos]{gfx_xpos()}} and \code{gfx_ypos()} return the X and Y coordinates, respectively,
of the mouse pointer when the last key or button was pressed.
}
